Change the Data type for INK to BIGINT

+++ migrations/1652503990197_create-ktp-result-table.js
@@ -17,7 +17,7 @@ exports.up = (pgm) => {
       notNull: true,
     },
     nik: {
-			type: 'INTEGER',
+			type: 'BIGINT',
 			unique: true,
 			notNull: true,
     },
